+++ pons/__init__.py
@@ -2,6 +2,7 @@ from . import abi
 from ._abi_types import ABIDecodingError
 from ._client import (
     Client,
+    ClientSession,
     RemoteError,
     ContractPanic,
     ContractLegacyError,
@@ -12,15 +13,27 @@ from ._client import (
 from ._contract_abi import (
     ContractABI,
     Constructor,
+    ConstructorCall,
     Method,
+    MethodCall,
     Event,
+    EventFilter,
     Error,
     Fallback,
     Receive,
     Either,
     Mutability,
 )
-from ._contract import CompiledContract, DeployedContract
+from ._contract import (
+    BoundConstructor,
+    BoundConstructorCall,
+    BoundMethod,
+    BoundMethodCall,
+    BoundEvent,
+    BoundEventFilter,
+    CompiledContract,
+    DeployedContract,
+)
 from ._entities import (
     Amount,
     Address,