+++ docs/changelog.rst
@@ -8,7 +8,8 @@ Changelog
 Added
 ^^^^^
 
-- `Client.transact()` takes an optional `return_events` argument, allowing one to get "return values" from the transaction via events. (PR_52_)
+- ``Client.transact()`` takes an optional ``return_events`` argument, allowing one to get "return values" from the transaction via events. (PR_52_)
+- Exposed ``ClientSession``, ``ConstructorCall``, ``MethodCall``, ``EventFilter``, ``BoundConstructor``, ``BoundConstructorCall``, ``BoundMethod``, ``BoundMethodCall``, ``BoundEvent``, ``BoundEventFilter`` from the top level. (PR_53_)
 
 
 Fixed
@@ -19,6 +20,7 @@ Fixed
 
 .. _PR_51: https://github.com/fjarri/pons/pull/51
 .. _PR_52: https://github.com/fjarri/pons/pull/52
+.. _PR_53: https://github.com/fjarri/pons/pull/53
 
 
 0.7.0 (09-07-2023)